requirements.txt
langchain
openai
python-dotenv
pypdf
pdfminer
pymongo
spacy